requirements.txt
aiohttp
azure-functions
beautifulsoup4
boto3
duckduckgo-search
faiss-cpu
huggingface_hub
langchain
openai
requests
sentence-transformers
wikipedia
https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp38-cp38-linux_x86_64.whl
# azure-functions
# boto3
# huggingface_hub
# langchain
# numpy
opencv-python-headless==4.5.3.56
pandas==1.5.3
# requests
scikit-learn==1.0.2
tflite-runtime